pyproject.toml
[tool.poetry]
name = "cashflow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
